+++ jgclark.Reviews/src/index.js
@@ -3,7 +3,7 @@
 //-----------------------------------------------------------------------------
 // Index for Reviews plugin
 // by Jonathan Clark
-// Last updated 2026-04-17 for v2.0.0.b20, @jgclark
+// Last updated 2026-04-20 for v2.0.0.b22, @jgclark
 //-----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
@@ -12,9 +12,9 @@ import { generateAllProjectsList } from './allProjectsListHelpers'
 import { renderProjectListsIfOpen } from './reviews'
 import { getReviewSettings } from './reviewHelpers'
 import { JSP, logDebug, logError, logInfo } from '@helpers/dev'
-import { pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
-import { showMessage } from '@helpers/userInput'
+import { backupSettings, pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
 
+export { getReviewSettings } from './reviewHelpers' // TODO(later): remove this export when we can stop testing settings issues
 export {
   finishReview,
   finishReviewAndStartNextReview,
@@ -62,7 +62,6 @@ const pluginID = 'jgclark.Reviews'
 
 export function init(): void {
   try {
-
     // Check for the latest version of the plugin, and if a minor update is available, install it and show a message. Do this in the background.
     DataStore.installOrUpdatePluginsByID([pluginJson['plugin.id']], false, false, false).then((r) =>
       pluginUpdated(pluginJson, r),
@@ -84,15 +83,14 @@ export async function onSettingsUpdated(): Promise<void> {
   // Only refresh the project list window if it is already open; do not open it from saving settings alone.
   try {
     const config = await getReviewSettings()
-    if (!config) throw new Error('No config found. Stopping.')
+    if (!config) throw new Error(`Can't get Review settings. Stopping.`)
     logDebug(pluginJson, 'Have updated Review settings; recalculating review list and refreshing project list UI if already open...')
     if (!(config.useDemoData ?? false)) {
       await generateAllProjectsList(config, true)
     }
     await renderProjectListsIfOpen(config)
   } catch (error) {
-    logError(pluginJson, JSP(error))
-    await showMessage(`Sorry, there's been an error updating the settings for this plugin: ${error.message}. Please check the logs for more details, and raise an issue on Discord.`, 'OK, thanks', 'Error updating settings')
+    logError(pluginJson, error.message)
   }
 }
 
@@ -101,6 +99,10 @@ export async function onUpdateOrInstall(): Promise<void> {
     logInfo(pluginID, `onUpdateOrInstall ...`)
     const updateSettingsResult = updateSettingData(pluginJson)
     logInfo(pluginID, `- updateSettingData code: ${updateSettingsResult}`)
+
+    // Backup the settings on all new installs (quietly)
+    // TODO: remove once issues around v2.0 settings have stopped
+    await backupSettings('jgclark.Reviews', `before_onUpdateOrInstall_v${pluginJson["plugin.version"]}`, true)
 
     // Tell user the plugin has been updated
     await pluginUpdated(pluginJson, { code: updateSettingsResult, message: 'unused?' })